requirements.txt
Django>=4.0
mysqlclient
django-cors-headers
djangorestframework-simplejwt
google-auth
requests
